@@ -1,32 +1,41 @@
 plugins {
     id 'com.android.application'
-    id 'kotlin-android'
+    id 'org.jetbrains.kotlin.android'
+    id 'com.google.gms.google-services'
+    id 'com.google.firebase.crashlytics'
 }
-apply plugin: 'com.google.gms.google-services'
-apply plugin: 'com.google.firebase.crashlytics'
+
 android {
-    compileSdkVersion 31
+    namespace "com.adityamhatre.bookingscheduler"
+    compileSdk 36
 
     defaultConfig {
         applicationId "com.adityamhatre.bookingscheduler"
-        minSdkVersion 26
-        targetSdkVersion 31
-        versionCode 31
-        versionName "2.4"
+        minSdk 26
+        targetSdk 36
+        versionCode 32
+        versionName "2.5"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
+    buildFeatures {
+        buildConfig true
+    }
+
+
     kotlinOptions {
         jvmTarget = '1.8'
     }
