@@ -27,10 +27,8 @@ import java.io.FileOutputStream
 import java.io.IOException
 import java.time.Instant
 import java.time.ZoneId
-import java.util.*
-
-
-
+import java.util.Date
+import java.util.Locale
 
 
 class ViewBookingDetails(private val bookingDetails: BookingDetails) : Fragment() {
@@ -218,5 +216,5 @@ private fun Instant.toHumanDate(): String {
 }
 
 private fun String.toTitleCase(): String {
-    return this[0].toUpperCase() + this.substring(1).toLowerCase(Locale.getDefault())
+    return this[0].uppercaseChar() + this.substring(1).lowercase(Locale.getDefault())
 }
